moar documentation; relocate some classes into new packages;
