+++ src/main/java/com/github/argon/sos/moreoptions/game/ui/ButtonMenu.java
@@ -1,6 +1,6 @@
 package com.github.argon.sos.moreoptions.game.ui;
 
-import com.github.argon.sos.moreoptions.util.UiUtil;
+import com.github.argon.sos.moreoptions.game.util.UiUtil;
 import lombok.Builder;
 import org.jetbrains.annotations.Nullable;
 import snake2d.util.gui.GuiSection;
