+++ src/main/java/com/github/argon/sos/moreoptions/game/ui/ColumnRow.java
@@ -3,7 +3,7 @@ package com.github.argon.sos.moreoptions.game.ui;
 import com.github.argon.sos.moreoptions.game.Action;
 import com.github.argon.sos.moreoptions.util.ClassUtil;
 import com.github.argon.sos.moreoptions.util.Lists;
-import com.github.argon.sos.moreoptions.util.UiUtil;
+import com.github.argon.sos.moreoptions.game.util.UiUtil;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
