+++ src/main/java/com/github/argon/sos/moreoptions/ui/panel/RacesPanel.java
@@ -7,7 +7,7 @@ import com.github.argon.sos.moreoptions.log.Logger;
 import com.github.argon.sos.moreoptions.log.Loggers;
 import com.github.argon.sos.moreoptions.util.Lists;
 import com.github.argon.sos.moreoptions.util.Maps;
-import com.github.argon.sos.moreoptions.util.UiUtil;
+import com.github.argon.sos.moreoptions.game.util.UiUtil;
 import init.race.Race;
 import init.sprite.UI.UI;
 import lombok.Builder;
